requirements.txt
discord.py
python-dotenv
aiosqlite
pytest
sqlalchemy
alembic
pytest-sqlalchemy